Added functionality for auto filling address
In LocationSearchInput the address is autofilled automatically if the parent component, Confirmation detects that is the case.

<Confirmation /> will pass address properties from the firestore to the <LocationSearchInput /> component

If LocationSearchInput detects these at any point then they will automatically fill up the input fields and disable the submit button.
This will also trigger the second step to connect discord accounts if needed